requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
Jinja2==3.1.4
psycopg2==2.9.9
requests==2.32.3
python-dotenv==1.0.1 